backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.29.0
aiosqlite==0.20.0
pydantic-settings==2.2.1
httpx==0.27.0
anthropic==0.25.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
